requirements.txt
atproto
google-generativeai
python-dotenv==1.0.0
schedule==1.2.0
requests==2.31.0
tweepy